make sure lmax is int

+++ pysm3/models/dust_realization.py
@@ -146,7 +146,7 @@ class ModifiedBlackBodyRealization(ModifiedBlackBody):
         )
 
         alm_small_scale = [
-            hp.almxfl(each, np.ones(min(synalm_lmax, 2.5 * self.nside)))
+            hp.almxfl(each, np.ones(int(min(synalm_lmax, 2.5 * self.nside))))
             for each in alm_small_scale
         ]
         map_small_scale = hp.alm2map(alm_small_scale, nside=self.nside)
@@ -204,7 +204,7 @@ class ModifiedBlackBodyRealization(ModifiedBlackBody):
             )
 
             alm_small_scale = hp.almxfl(
-                alm_small_scale, np.ones(min(2.5 * self.nside, synalm_lmax + 1))
+                alm_small_scale, np.ones(int(min(2.5 * self.nside, synalm_lmax + 1)))
             )
             output[key] = hp.alm2map(alm_small_scale, nside=self.nside) * output_unit
             output[key] *= modulate_map_I
